+++ bot/callback-queries/more/index.ts
@@ -1,0 +1,3 @@
+import { more } from '@bot/general-requests/more';
+
+export default ['more', more];
